add example usage of hash map and make structs opaque
